improve section parsing
